+++ geopackage-ios/GPKGBoundingBox.m
@@ -141,19 +141,19 @@
     return size;
 }
 
--(GPKGBoundingBox *) complementaryWithMaxLongitude: (double) maxLongitude{
+-(GPKGBoundingBox *) complementaryWithMaxLongitude: (double) maxProjectionLongitude{
     
     GPKGBoundingBox * complementary = nil;
     
     NSDecimalNumber *adjust = nil;
     
-    if([self.maxLongitude doubleValue] > maxLongitude){
-        if([self.minLongitude doubleValue] >= -maxLongitude){
-            adjust = [[NSDecimalNumber alloc] initWithDouble:-2 * maxLongitude];
+    if([self.maxLongitude doubleValue] > maxProjectionLongitude){
+        if([self.minLongitude doubleValue] >= -maxProjectionLongitude){
+            adjust = [[NSDecimalNumber alloc] initWithDouble:-2 * maxProjectionLongitude];
         }
-    }else if([self.minLongitude doubleValue] < -maxLongitude){
-        if([self.maxLongitude doubleValue] <= maxLongitude){
-            adjust = [[NSDecimalNumber alloc] initWithDouble:2 * maxLongitude];
+    }else if([self.minLongitude doubleValue] < -maxProjectionLongitude){
+        if([self.maxLongitude doubleValue] <= maxProjectionLongitude){
+            adjust = [[NSDecimalNumber alloc] initWithDouble:2 * maxProjectionLongitude];
         }
     }
     
@@ -172,6 +172,53 @@
 
 -(GPKGBoundingBox *) complementaryWebMercator{
     return [self complementaryWithMaxLongitude:PROJ_WEB_MERCATOR_HALF_WORLD_WIDTH];
+}
+
+-(GPKGBoundingBox *) boundCoordinatesWithMaxLongitude: (double) maxProjectionLongitude{
+
+    GPKGBoundingBox *bounded = [[GPKGBoundingBox alloc] initWithBoundingBox:self];
+    
+    double minLongitude = fmod([self.minLongitude doubleValue] + maxProjectionLongitude, 2 * maxProjectionLongitude)
+        - maxProjectionLongitude;
+    double maxLongitude = fmod([self.maxLongitude doubleValue] + maxProjectionLongitude, 2 * maxProjectionLongitude)
+        - maxProjectionLongitude;
+
+    [bounded setMinLongitude:[[NSDecimalNumber alloc] initWithDouble:minLongitude]];
+    [bounded setMaxLongitude:[[NSDecimalNumber alloc] initWithDouble:maxLongitude]];
+
+    return bounded;
+}
+
+-(GPKGBoundingBox *) boundWgs84Coordinates{
+    return [self boundCoordinatesWithMaxLongitude:PROJ_WGS84_HALF_WORLD_LON_WIDTH];
+}
+
+-(GPKGBoundingBox *) boundWebMercatorCoordinates{
+    return [self boundCoordinatesWithMaxLongitude:PROJ_WEB_MERCATOR_HALF_WORLD_WIDTH];
+}
+
+-(GPKGBoundingBox *) expandCoordinatesWithMaxLongitude: (double) maxProjectionLongitude{
+    
+    GPKGBoundingBox *expanded = [[GPKGBoundingBox alloc] initWithBoundingBox:self];
+    
+    double minLongitude = [self.minLongitude doubleValue];
+    double maxLongitude = [self.maxLongitude doubleValue];
+    
+    if (minLongitude > maxLongitude) {
+        int worldWraps = 1 + (int) ((minLongitude - maxLongitude) / (2 * maxProjectionLongitude));
+        maxLongitude += (worldWraps * 2 * maxProjectionLongitude);
+        [expanded setMaxLongitude:[[NSDecimalNumber alloc] initWithDouble:maxLongitude]];
+    }
+    
+    return expanded;
+}
+
+-(GPKGBoundingBox *) expandWgs84Coordinates{
+    return [self expandCoordinatesWithMaxLongitude:PROJ_WGS84_HALF_WORLD_LON_WIDTH];
+}
+
+-(GPKGBoundingBox *) expandWebMercatorCoordinates{
+    return [self expandCoordinatesWithMaxLongitude:PROJ_WEB_MERCATOR_HALF_WORLD_WIDTH];
 }
 
 @end